core: move IoReader and IoWriter into utils mod
Refactor the `IoReader`/`IoWriter` into `core::utils`, so that it can be re-used
by external crates.

`IoWrite` was renamed to `IoWriter` for consistency.
